Remove unused utils
