+++ codesign/utils/Math.ts
@@ -1,7 +1,3 @@
-export function getRandomFloat(min: number, max: number): number {
-  return Math.random() * (max - min) + min;
-}
-
 /**
  *
  * @param min inclusive minimum value
